feat: use dynamo db (#1220)

+++ src/job/file_list.rs
@@ -72,19 +72,20 @@ async fn move_file_list_to_storage() -> Result<(), anyhow::Error> {
             continue;
         }
         log::info!("[JOB] convert file_list: {}", file);
-
-        match upload_file(&local_file, &file_name).await {
-            Ok(_) => match fs::remove_file(&local_file) {
-                Ok(_) => {}
-                Err(e) => {
-                    log::error!(
-                        "[JOB] Failed to remove file_list {}: {}",
-                        local_file,
-                        e.to_string()
-                    )
-                }
-            },
-            Err(e) => log::error!("[JOB] Error while uploading file_list to storage {}", e),
+        if !CONFIG.common.use_dynamo_meta_store {
+            match upload_file(&local_file, &file_name).await {
+                Ok(_) => match fs::remove_file(&local_file) {
+                    Ok(_) => {}
+                    Err(e) => {
+                        log::error!(
+                            "[JOB] Failed to remove file_list {}: {}",
+                            local_file,
+                            e.to_string()
+                        )
+                    }
+                },
+                Err(e) => log::error!("[JOB] Error while uploading file_list to storage {}", e),
+            }
         }
     }
     Ok(())
